add cli mod
